Fixed more review comments and codacy issues
