+++ ha/core/system_health/model/health_event.py
@@ -20,6 +20,8 @@ class HealthEvent:
     required by System Health module for updating the health.
     """
 
+    VERSION = "v1"
+
     def __init__(self, event_id: str, event_type: str, severity: str, site_id: int, rack_id: int, cluster_id: str, storageset_id: int,
                  node_id: int, host_id: str, resource_type: str, timestamp: str, resource_id: str, specific_info: dict=None):
         """
